requirements.txt
flask==3.0.3
pdfplumber==0.11.4
httpx==0.27.2
python-dotenv==1.0.1
